+++ SCT2/trunk/plugins/org.yakindu.sct.model.stext/src-gen/org/yakindu/sct/model/stext/stext/TimeEventSpec.java
@@ -2,6 +2,7 @@
  * <copyright>
  * </copyright>
  *
+
  */
 package org.yakindu.sct.model.stext.stext;
 
@@ -64,12 +65,12 @@ public interface TimeEventSpec extends EventSpec
    * </p>
    * <!-- end-user-doc -->
    * @return the value of the '<em>Value</em>' attribute.
-   * @see #setValue(int)
+   * @see #setValue(long)
    * @see org.yakindu.sct.model.stext.stext.StextPackage#getTimeEventSpec_Value()
    * @model
    * @generated
    */
-  int getValue();
+  long getValue();
 
   /**
    * Sets the value of the '{@link org.yakindu.sct.model.stext.stext.TimeEventSpec#getValue <em>Value</em>}' attribute.
@@ -79,7 +80,7 @@ public interface TimeEventSpec extends EventSpec
    * @see #getValue()
    * @generated
    */
-  void setValue(int value);
+  void setValue(long value);
 
   /**
    * Returns the value of the '<em><b>Unit</b></em>' attribute.
